+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog for the [`ghc-typelits-extra`](http://hackage.haskell.org/package/ghc-typelits-extra) package
 
 # Unreleased
+* Do not require equality to unit constraint for wanted InEqs for `CLog` and `CLogWZ`.
 * Fix wrong reification from `CLogWZ` to `CLog` causing GHC panic. [#71](https://github.com/clash-lang/ghc-typelits-extra/issues/71)
 
 # 0.5.2 *December 3rd 2025*